downloader.py
import os
import time
import requests
from io import BytesIO
from PIL import Image

# Configuration
ASSET_DIR = "assets/faces"
TARGET_COUNT = 50
IMAGE_SIZE = (512, 512)

def clean_directory():
    os.makedirs(ASSET_DIR, exist_ok=True)
    print(f"📁 Verified directory: {ASSET_DIR}")

def download_faces():
    clean_directory()
    
    print("🚀 Initiating high-quality StyleGAN face dataset download...")
    print("Downloading from public GAN repository (thispersondoesnotexist.com or fallbacks)")
    
    session = requests.Session()
    session.headers.update({
        'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36'
    })

    success_count = 0
    attempts = 0
    max_attempts = TARGET_COUNT * 3
    
    # Simple fallback faces (Unsplash Source API or similar, although Unsplash source is deprecated)
    # TPDNE is the best for GAN faces. If it fails, we use a robohash or dicebear as *last* resort, 
    # but TPDNE usually handles sequential requests if we rate-limit to 1-2 seconds.
    
    while success_count < TARGET_COUNT and attempts < max_attempts:
        attempts += 1
        
        try:
            # Add timestamp to bust caches and ensure unique face every time
            url = f"https://thispersondoesnotexist.com/?req={int(time.time() * 1000)}"
            response = session.get(url, timeout=10)
            
            if response.status_code == 200:
                img = Image.open(BytesIO(response.content)).convert("RGB")
                img = img.resize(IMAGE_SIZE, Image.Resampling.LANCZOS)
                
                filename = f"{success_count:03d}.png"
                filepath = os.path.join(ASSET_DIR, filename)
                
                img.save(filepath, "PNG")
                success_count += 1
                
                print(f"✅ Downloaded {success_count}/{TARGET_COUNT}: {filename}")
                
                # Polite delay to prevent rate-limiting
                time.sleep(1.5)
            else:
                print(f"⚠️ Warning: Got status code {response.status_code}. Retrying...")
                time.sleep(3)
                
        except Exception as e:
            print(f"❌ Error on attempt {attempts}: {e}")
            time.sleep(5)
            
    if success_count == TARGET_COUNT:
        print("\n🎉 SUCCESS! All 50 high-quality AI generated faces downloaded and formatted.")
    else:
        print(f"\n⚠️ Reached maximum attempts. Downloaded {success_count}/{TARGET_COUNT} faces.")
        print("Demo will gracefully loop over available faces.")

if __name__ == "__main__":
    download_faces()
