requirements.txt
gradio>=4.40.0
Pillow>=11.0.0
numpy>=1.26.0
requests>=2.31.0
